[TG-6412][TG-6489] Productionize "Simplify basket bar": round 3 (#5230)
* [TG-6412] move OpenBoxButton one level up

In preparation to unifying the Mobile/Desktop banner content. There won't be
separate Desktop and Mobile folders, so this component will live alongside the
main component.

* [TG-6412] move ExpandBoxSummaryButton one level up

The same reason as previously: there won't be Desktop&Mobile folders, instead
this component will be included from the main banner component.

* [TG-6412] productionize for OpenBoxButton

* [TG-6412] assume isSimplifyBasketBarEnabled=true for MobileBanner

* [TG-6412] assume isSimplifyBasketBarEnabled=true for Desktop banner

* [TG-6412] move all prop types to a single place

* [TG-6412] Inline mobile banner before unifying

* [TG-6412] Inline desktop banner before unifying

* [TG-6412] remove now-unused Desktop and Mobile components

* [TG-6412] remove unused styles

* [TG-6412] unify styles of the main boxSummaryBanner

* [TG-6412] introduce JS media query; replace hard-coded view with var

* [TG-6412] unify styles of buttonContainer

* [TG-6412] replace colors and typography of error message with modern design system

This is a slight change of functionality (the colors don't correspond 1-1), but
I feel it's warranted because it brings an old component more in line with the
current look guidelines.

* [TG-6412] unify desktop/mobile bar

* [TG-6412] replace all bar heights with basketBarHeight

* [TG-6412] expand on tap only on mobile

* [TG-6412] remove unused variables in BoxSummaryBanner

* [TG-6412] remove extracting and passing now-unused props

* [TG-6412] remove flag in BoxSummaryBanner

* [TG-6412] rename selector to createGetPromoStoreEntry for clarity

* [TG-6412] move selectors to their proper places

* [TG-6412] remove experiment mentions from BoxSummaryOverlay

* [TG-6412] remove experiment mentions from ActionBar

* [TG-6412] move CheckoutCounter to a proper place

BannerButton will be replaced, so its part is being moved to a place accessible
from both old and the new component.

* [TG-6412] replace BannerButton with a component using citrus/Button

* [TG-6412] define spacing symmetrically; reduce button padding for small screens

* [TG-6412] remove RecipeList (that was accessed from Banner)

* [TG-6412] remove feature flag in ExpandBoxSummaryButton

* [TG-6412] replace ExpandBoxSummaryButton with citrus/Button

* [TG-6412] set flag to true in Title

* [TG-6412] set flag to true in Price

* [TG-6412] set flag to true in Description

* [TG-6412] move ExpandBoxSummaryButton's Contents to new PriceAndDiscountTip

This will be a component that contains all logic for displaying the
information (instead of the logic being spread across Title, Description,
Price); and ExpandBoxSummaryButton will change so that it's actually a button,
and the "no button on mobile" logic will be moved to its parent.

* [TG-6412] remove unused variables in PriceAndDiscountTip and ExpandBoxSummaryButton

* [TG-6412] move including standalone PriceAndDiscountTip to parent

Also move logic "don't show if pop-over is shown" to the button component.

* [TG-6412] inline Title

As it's only used in one place, it doesn't make sense to have a separate file
high up in the hierarchy.  Also some logic is repeated in Title&Description,
and it'll be clearer when they're both in the same place.

* [TG-6412] inline Description

* [TG-6412] clarify three-case condition; better and unified class names

* [TG-6412] fix retrieving promo code information for new users

It used the data dependency on the `promoCode` value of the result of the
`usePricing` hook; but this is incorrect because when there are not enough
recipes to checkout, that result is empty.  On the other hand, the basket slice
contains the promo code to apply even if there are no recipes; so we
cross-check the basket's promoCode with promoStore entry for that promo code.

It also is the same selector that's used for checking whether we should
`promoGet()` - retrieve the promoStore entry for that code - in

gousto-webclient/src/apps/webclient/src/routes/Menu/components/BoxSummary/Banner/BoxSummaryBanner.js:40

* [TG-6412] rename Price props to align with pricing response

* [TG-6412] replace loading dash look and feel; pass aggregate prop instead of separate

While it is not requested by design, the dash was present in experiment
development, and there is no better way to show the loading state.  (The best
would be to show the previous values of the price, but a naive `usePrevious`
usage didn't work out.)

* [TG-6412] move Price to the folder it's used from

It could have been inlined just like Title and Description; and that would
uncover the parallel logic with `isDiscountEnabled`; however the loading state
gets in the way, and the early return is clearer than multi-level ternary
conditionals.

* [TG-6412] remove remaining mentions of Simplify Basket Bar

* [TG-6412] fix some types and unit tests

* [TG-6412] reformat files to avoid merge conflicts

* [TG-6412] delete unit tests for deleted components

* [TG-6412] fix some eslint warnings

* [TG-6412] fix Price unit test

* [TG-6412] update unit test for BoxSummaryBanner

* [TG-6412] update unit test for BoxSummaryOverlayDesktop

* [TG-6412] fix some more eslint and type warnings

* [TG-6412] add unit tests for status selectors; lessen duplication in some of them

* [TG-6412] better prop types for CheckoutButton and PriceAndDiscountTip

* [TG-6412] fix alignment in the PriceAndDiscountTip

* [TG-6412] refactor utilHooks; fix tip for existing users & no promo code

- remove unused properties from the type (used only in Price, which now gets
them directly from the usePricing hook response)
- unify two types because they now had the same essential data
- rename the type to DiscountDescriptor for clarity
- useDiscountTip change to formatDiscountTip because a pure function is
preferable to a hook

And the impetus for the change: there was a bug where if you sign up without a
promo code then go to Choose recipes, you see the tip as "£0 off youu box".
This is due to an incorrect API response that got into the pricing hook
response; which claims "flatDiscountApplied=true" with the amount=0.  Added
defensive checks around the actual discount amounts, so the UI considers a
discount enabled only when the amount is a string representing a positive
number.

* [TG-6412] move and rename utilHooks -> priceAndDiscountTipUtils

Since the only user of these utils is the PriceAndDiscountTip component, move
here.

Since it's not only hooks now, remove "hooks" from the name.

To better reflect whose utils it is, prepend "priceAndDiscountTip".

* [TG-6412] fix priceAndDiscountTipUtils test

* [TG-6412] update e2e tests

* [TG-6412] CR: Use Pricing instead of Pick<Pricing, 'necessary-props'>

* [TG-6412] extract&move the logic to ensure promo code entry is retrieved

Extract from the main banner component; move it to be a part of the hook that
uses that information.

* [TG-6412] ExpandBoxSummaryButton: use `useSelector` instead of mapStateToProps

* [TG-6412] avoid mocking useSelector in useDiscountDescriptor tests

* [TG-6412] fix typecheck:tests

* [TG-6412] fix eslint warnings

* [TG-6412] clarify step name: log out *then* log in

* [TG-6412] change test tag to allow launching single test files

`yarn install --tag criticalUserJourney` should launch just one test, not two.

* [TG-6412] make isDisabled dependent on previous logic for isDisabled||isPending

* [TG-6412] fix eslint warning

For some reason, it showed in Github actions but didn't locally.