FEF-489: Tolerate ignored files in lint (#5280)
FEF-489

This PR fixes a couple of problems with the workspace-changes plugin setup:

- Our code health scripts can't tolerate warnings caused by us passing ignored files
- We don't identify when ESLint has to be re-run on everything

I did think about parsing the eslintignore file directly, but this didn't seem straightforward, and in any case the immediate problem seemed to be in the code health scripts, which we might be replacing at some point anyway.

What I have done, however, is refactor workspace-changes into a monorepo-build plugin, that gets some of its sources from an internal library. The Yarn plugin binds to the CLI and prepares a 'context', which the library function can call. We can now use this approach to unit test the monorepo build plugin, which I think should give us a lot more confidence going forward.

I also altered the code health script - we now disregard instances where a file has a warning due to an ignore pattern.