[TG-6354] Show specific error on signup login failed: round 2 (#5342)
* [TG-6354] Alert now has severity prop to display error or success

Prop naming cribbed from https://mui.com/material-ui/react-alert/ .

* [TG-6354] capture the signup login error correctly

This is the meat of the bug fix.  `checkoutPostSignup` dispatches `login`
inside of a `try` block; but `login` has its own `try` around the actual API
call; which is not re-thrown inside the catch.  This means that the execution
flow never reached the `catch` in `checkoutPostSignup` (this is corroborated by
checking the log for April - no instances of the FE logging even though we know
from CC requests that users ran into such a situation).

* [TG-6354] display the situation as a success-mode Alert

The user's account has been created, only the login call didn't succeed: so we
can reframe it as success and ask the user to re-try the login.

* [TG-6354] do not show the main signup CTA in case of signup login error

* [TG-6354] fix eslint errors

* [TG-6354] fix unit tests