FEF-511 / Optionally run main Jest tests in GHA (#5362)
This PR implements FEF-511

As we move to modules, there will be builds going through where we don't need to run the Jest tests for the main webclient. The 'rules' for testing modules we've decided are that we

    run unit tests for the changed modules
    run integration tests for their dependents
    run e2e for the whole app

# Implementation

I've created a new monorepo build command very quickly that checks whether a module requires unit tests. The check is simple: does the module have any changed files? If so, assume we need to re-run tests. We could refine that later if required.

Arguably there are cases where a file outside the webclient module could change the results of a test run. For example:

    a change to a library used to help configure the test runner
    a change to a common config file like tsconfig.json

We could possibly have a way of tracking these dependencies... or we could accept a tradeoff that anyone changing those kinds of files, should manually run all the tests in the project. They would want to manually test things like the dev flows anyway.

My feeling is that this is an acceptable tradeoff, and we can refine the plugin later... but perhaps others feel differently.

# Command code

There's a little repetition in the command boilerplate. I'm waiting for the "rule of three" before refactoring that, however.
Testing

I've tested that this change still runs Jest tests when a file changes in the main webclient module. However, otherwise it will skip the unit test job - that saves about eleven or so minutes. For changes to other feature modules, or changes to documentation, this will represent a large saving. So I'm hopeful we can merge this.