[TG-5004] fix default export warnings in Radishes domain (#5436)
* [TG-5004] actions/filters

* [TG-5004] components/Carousel

* [TG-5004] components/Icon

* [TG-5004] components/Item

* [TG-5004] components/Modal

* [TG-5004] components/Receipt

* [TG-5004] optimizelyRolloutsConfig

* [TG-5004] recipesConfig

* [TG-5004] MainLayout

Note: it uses explicit aliasing just because so many of the importing files use
this name.  In the ideal world we'd have the MainLayoutContainer name in all
those places.

* [TG-5004] boxSummaryReducers

* [TG-5004] collectionsReducers

* [TG-5004] filtersReducers

* [TG-5004] menuReducers

* [TG-5004] recipesReducers

* [TG-5004] Menu

* [TG-5004] BoxSummary

* [TG-5004] BoxSummaryContainer

* [TG-5004] DetailCarousel

Renamed to make it clear it's not a Carousel from components/Carousel.  In the
ideal world the whole directory and file name would be

src/routes/Menu/components/Detail/DetailCarousel/DetailCarousel.test.js

for less confusion.

* [TG-5004] InfoToggle

* [TG-5004] MenuRoute

* [TG-5004] browserHelperUtils