[TG-6608] add logs to analyze Recaptcha errors on signup login (#5440)
It seems from the correlation by timestamps between the server logs and
feLogging logs that all remaining Recaptcha errors on signup login attempts are
with the same error code: `timeout-or-duplicate`.  To better understand why
they happen, more logs were added:

- when Start your subscription button is clicked (= submitCheckingRecaptcha)
- when Recaptcha token is ready or expires (handleRecaptchaChange)
- correlation data to the server logs - to definitely link the two events

So if the timestamps between handleRecaptchaChange and the server errors is
>2min, we know for sure it's a timeout issue.  If there is no
handleRecaptchaChange(null) in between, then we know the library doesn't
function as it claims (it claims it does this call to show expiration on client
side).  If there is such a call, we know that the client can potentially catch
it but something else prevents us re-generating the token.