[TG-6657] rename stepName to secondarySlug
First it's not stepName, it's a slug because it's the part of the URL in
<server>/signup/<thisthing>.  And the steps in a config already have a thing
called name, so it is confusing to use the same name for a different thing.

Additionally, it's not just a step slug, because nowadays we have
stand-alone (not in the wizard's 1-2-3 steps) pages: Sell the proposition,
Check account, etc.  Therefore it's clearer to use a more generic name for this
thing - the idea is "signup" is the primary slug, and "about" or "postcode"
would be the secondary slug.