[TG-6657] attempt to fix Optimizely allocation for wizard experiments (#5503)
* [TG-6657] rename sellThePropositionPagePath -> sellThePropositionPageSlug

* [TG-6657] rename stepName to secondarySlug

First it's not stepName, it's a slug because it's the part of the URL in
<server>/signup/<thisthing>.  And the steps in a config already have a thing
called name, so it is confusing to use the same name for a different thing.

Additionally, it's not just a step slug, because nowadays we have
stand-alone (not in the wizard's 1-2-3 steps) pages: Sell the proposition,
Check account, etc.  Therefore it's clearer to use a more generic name for this
thing - the idea is "signup" is the primary slug, and "about" or "postcode"
would be the secondary slug.

* [TG-6657] change syntax for clarity

Immutable.fromJS() is recursive, so putting in an additional `Immutable.List`
seems to be doing someting it doesn't.

* [TG-6657] remove signup wizard steps from persistence

The signup steps will be retrieved based on the feature flags, checked at the
moment of landing, so having an additional way of overwriting them from old
cookies is not safe.

* [TG-6657] rename steps -> stepNames for clarity

There are cases where "step" represents an Immutable.Map of keys name and slug;
and cases where "step" means just a step name.  This is move towards separation
of those two things.

* [TG-6657] load steps and perform redirect to proper step in different stages

* [TG-6657] rename firstStepUrl to firstStepSlug

Since it's just a part of a URL, calling it "slug" is clearer.

* [TG-6657] fix unit tests

* [TG-6657] rename function and variable names to match intent

The result of calling a factory is a *function*, so it starts with `get`.  The
word "Experiment" in a `isFooEnabled` construct is redundant.

* [TG-6657] add a unit test for missing 1% of coverage

* [TG-6657] remove temporary code

* [TG-6657] fix defaultProps

* [TG-6657] fix the old test that started failing